rename folder, create BankConnection component
